+++ react-native-scripts/src/util/packager.js
@@ -1,7 +1,14 @@
 // @flow
 
-import { PackagerLogsStream, Project, ProjectSettings, ProjectUtils } from 'xdl';
+import {
+  PackagerLogsStream,
+  Project,
+  ProjectSettings,
+  ProjectUtils,
+} from 'xdl';
 
+import _ from 'lodash';
+import spawn from 'cross-spawn';
 import ProgressBar from 'progress';
 import bunyan from '@expo/bunyan';
 import chalk from 'chalk';
@@ -39,7 +46,9 @@ async function cleanUpPackager(projectDir) {
   if (result === 'stopFailed') {
     // find RN packager pid, attempt to kill manually
     try {
-      const { packagerPid } = await ProjectSettings.readPackagerInfoAsync(projectDir);
+      const { packagerPid } = await ProjectSettings.readPackagerInfoAsync(
+        projectDir
+      );
       process.kill(packagerPid);
     } catch (e) {
       process.exit(1);
@@ -54,12 +63,57 @@ function shouldIgnoreMsg(msg) {
     msg.indexOf('Warning: PropTypes has been moved to a separate package') >= 0;
 }
 
-function run(onReady: () => ?any, options: Object = {}, isInteractive = false) {
+function run(
+  onReady: () => ?any,
+  options: Object = {},
+  isInteractive?: boolean = false
+) {
   let packagerReady = false;
   let needsClear = false;
   let logBuffer = '';
   let progressBar;
   const projectDir = process.cwd();
+
+  if (process.platform !== 'win32') {
+    const watchmanExists = spawn.sync('which', ['watchman']).status === 0;
+
+    if (process.platform === 'darwin' && !watchmanExists) {
+      const watcherDetails = spawn
+        .sync('sysctl', ['kern.maxfiles'])
+        .stdout.toString();
+      if (parseInt(watcherDetails.split(':')[1].trim()) < 5242880) {
+        log.withTimestamp(
+          `${chalk.red(`Unable to start server`)}
+See https://git.io/v5vcn for more information, either install watchman or run the following snippet:
+${chalk.cyan(`  sudo sysctl -w kern.maxfiles=5242880
+  sudo sysctl -w kern.maxfilesperproc=524288`)}
+        `
+        );
+        process.exit(1);
+      }
+    } else if (!watchmanExists) {
+      try {
+        const watcherDetails = spawn
+          .sync('sysctl', ['fs.inotify.max_user_watches'])
+          .stdout.toString();
+        if (parseInt(watcherDetails.split('=')[1].trim()) < 12288) {
+          log.withTimestamp(
+            `${chalk.red(`Unable to start server`)}
+  See https://git.io/v5vcn for more information, either install watchman or run the following snippet:
+  ${chalk.cyan(`  sudo sysctl -w fs.inotify.max_user_instances=1024
+    sudo sysctl -w fs.inotify.max_user_watches=12288`)}`
+          );
+          process.exit(1);
+        }
+      } catch (e) {
+        // note(brentvatne): I'm not sure why stdout is null for some OS's
+        // https://github.com/react-community/create-react-native-app/issues/391
+        log.withTimestamp(
+          'Warning: Unable to run `sysctl fs.inotify.max_user_watches`. If you encounter issues, please refer to https://git.io/v5vcn'
+        );
+      }
+    }
+  }
 
   const handleLogChunk = chunk => {
     // pig, meet lipstick
@@ -93,13 +147,9 @@ function run(onReady: () => ?any, options: Object = {}, isInteractive = false) {
 
     if (packagerReady) {
       const message = `${chunk.msg.trim()}\n`;
-      if (chunk.level <= bunyan.INFO) {
-        log.withTimestamp(message);
-      } else if (chunk.level === bunyan.WARN) {
-        log.withTimestamp(chalk.yellow(message));
-      } else {
-        log.withTimestamp(chalk.red(message));
+      logWithLevel(chunk);
 
+      if (chunk.level === bunyan.ERROR) {
         // if you run into a syntax error then we should clear log output on reload
         needsClear = message.indexOf('SyntaxError') >= 0;
       }
@@ -119,12 +169,15 @@ function run(onReady: () => ?any, options: Object = {}, isInteractive = false) {
   let packagerLogsStream = new PackagerLogsStream({
     projectRoot: projectDir,
     onStartBuildBundle: () => {
-      progressBar = new ProgressBar('Building JavaScript bundle [:bar] :percent', {
-        total: 100,
-        clear: true,
-        complete: '=',
-        incomplete: ' ',
-      });
+      progressBar = new ProgressBar(
+        'Building JavaScript bundle [:bar] :percent',
+        {
+          total: 100,
+          clear: true,
+          complete: '=',
+          incomplete: ' ',
+        }
+      );
 
       log.setBundleProgressBar(progressBar);
     },
@@ -146,7 +199,9 @@ function run(onReady: () => ?any, options: Object = {}, isInteractive = false) {
           log.withTimestamp(chalk.red(`Failed building JavaScript bundle`));
         } else {
           let duration = endTime - startTime;
-          log.withTimestamp(chalk.green(`Finished building JavaScript bundle in ${duration}ms`));
+          log.withTimestamp(
+            chalk.green(`Finished building JavaScript bundle in ${duration}ms`)
+          );
         }
       }
     },
@@ -193,5 +248,85 @@ function run(onReady: () => ?any, options: Object = {}, isInteractive = false) {
     }
   );
 }
+
+const logStackTrace = (chunk, logFn, nestedLogFn, colorFn) => {
+  let traceInfo;
+  try {
+    traceInfo = JSON.parse(chunk.msg);
+  } catch (e) {
+    return logFn(colorFn(chunk.msg));
+  }
+
+  let { message, stack } = traceInfo;
+  logFn(colorFn(chalk.bold(message)));
+
+  const isLibraryFrame = line => {
+    return line.startsWith('node_modules');
+  };
+
+  let stackFrames = _.compact(stack.split('\n'));
+  let lastAppCodeFrameIndex = _.findLastIndex(stackFrames, line => {
+    return !isLibraryFrame(line);
+  });
+  let lastFrameIndexToLog = Math.min(
+    stackFrames.length - 1,
+    lastAppCodeFrameIndex + 2 // show max two more frames after last app code frame
+  );
+  let unloggedFrames = stackFrames.length - lastFrameIndexToLog;
+
+  // If we're only going to exclude one frame, just log them all
+  if (unloggedFrames === 1) {
+    lastFrameIndexToLog = stackFrames.length - 1;
+    unloggedFrames = 0;
+  }
+
+  for (let i = 0; i <= lastFrameIndexToLog; i++) {
+    let line = stackFrames[i];
+    if (!line) {
+      continue;
+    } else if (line.match(/react-native\/.*YellowBox.js/)) {
+      continue;
+    }
+
+    if (line.startsWith('node_modules')) {
+      nestedLogFn(colorFn('- ' + line));
+    } else {
+      nestedLogFn(colorFn('* ' + line));
+    }
+  }
+
+  if (unloggedFrames > 0) {
+    nestedLogFn(
+      colorFn(
+        `- ... ${unloggedFrames} more stack frames from framework internals`
+      )
+    );
+  }
+};
+
+const logWithLevel = chunk => {
+  if (!chunk.msg) {
+    return;
+  }
+
+  let colorFn = (str) => str;
+  if (chunk.level === bunyan.WARN) {
+    colorFn = chalk.yellow;
+  } else if (chunk.level === bunyan.ERROR) {
+    colorFn = chalk.red;
+  }
+
+  if (chunk.includesStack) {
+    logStackTrace(chunk, log.withTimestamp, log, colorFn);
+  } else {
+    logLines(chunk.msg, log.withTimestamp, colorFn);
+  }
+};
+
+const logLines = (msg, logFn, colorFn) => {
+  for (let line of msg.split('\n')) {
+    logFn(colorFn(line));
+  }
+};
 
 export default { run };